Fix C++11 exception handling in destructors
In C++11 default behavior when throwing an exception in destructors has
changed. Default for all destructors is noexcept, which means the
program will be immediately terminated when an exception in a destructor
is thrown.  To change this behavior to allow throwing catchable
exceptions in a destructor we add the noexcept(false) specifier.